+++ src/app/_components/sections/projects/ProjectDesktopCard.tsx
@@ -1,7 +1,7 @@
 'use client'
-import { MAIN_PROJECT_DATA } from 'src/interfaces/common'
 import { FillImage, IconButton, Link$, useVisibility } from '~/components/common'
 import { Arrow } from '~/src/components/icons'
+import type { MAIN_PROJECT_DATA } from '~/src/interfaces/Common'
 
 interface ProjectCardDesktopProps {
     isReverse: boolean
