+++ src/app/api/main/getProjects/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { API_MAIN_PROJECTS, API_RESPONSE, MAIN_PROJECT_DATA } from '~/interfaces/Common'
+import { API_MAIN_PROJECTS, API_RESPONSE, MAIN_PROJECT_DATA } from '~/src/interfaces/Common'
 import { getMainProjectDB } from '~/utils/NotionUtil'
 
 export async function GET(_: NextRequest) {
