+++ src/app/api/main/getTimelines/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { API_MAIN_TIMELINES, API_RESPONSE, MAIN_TIMELINE_DATA } from '~/interfaces/Common'
+import { API_MAIN_TIMELINES, API_RESPONSE } from '~/src/interfaces/Common'
 import { getMainTimelineDB } from '~/utils/NotionUtil'
 
 export async function GET(_: NextRequest) {
